Bugfix: Commented lmake_include is not ignored (#35)
When an lmake_include is commented it should be ignored and not tried to be included but it is.

When file was preprocessed, comment lines were not ignored, this meant that the `lmake_include` command still included the file. Now, when the line starts with `--` the line is skipped from preprocessing.