Enhancement: verbose flag (#37)
LMake should have a verbose mode where it prints the exact commands executing instead of a simple message saying `[+] Compiling ...`. This means that `lmake_exec` should only print the command name and print the entire command when verbose is activated. 

This should be implemented as a flag, `--verbose`. As it's done with `--recompile`.

This is implemented by the settings inside `lmake_data.settings`, when this is on, the whole command is printed; when not, a shorter and prettier message is printed.