Bugfix: crash when unexistent source file tries to be compiled (#34)
Bugfix for issue #34.

When an unexistent file is set to be compiled a crash occurs, this might happen because of the date comparison that is done in order to know if the file needs to be compiled or not.

Fixed by checking if the source file exists before comparing dates. Object file was checked but source file not by some reason.