Enhancement: Use path env variable when executing processes. (#47)
For creating more general lmake scripts that could potentially be more cross platform and more readable, when setting the compiler, linker or executing a system process, instead of giving the full path, the user should be able to just add the program name and lmake search on the `PATH` variable.

On function `run_process` there is a check if a path is there. If there the specified process is inside a path just the process is executed, if not, the executable is searched on `PATH`. This functionality is similar to what bash does on the command line.