Enhancement: more deep tests (#68)
Deeper tests have been added to ensure basic functions at least don't break. CHVAR test has been improved with a new case: a variable does not exist and the value is requested to change. Apart from that, a test for regex has been added that tests when all is correct, when no regex has been added and when a path does not exist.