Bug: malloc error (#84)
Error was caused because of out of bounds on `std::shared_ptr` that is returned with function `os::read_file`. `length + 1` was being read when the buffer was of size `length`.

Some buffers probably work depending on the alignment that the standard library used.

This was tested on old branch where lmake was trying to be built with lmake.