Bug: compilation times (#89)
The compilation times before this patch where 45 seconds on my laptop. After this patch, compilation times went down to 10 seconds on the same machine.

The problem was that `spdlog` was used as a header only library instead of using the static library compiled. By defining `SPDLOG_COMPILED_LIB` the header removes all the implementation.